move view-config file
