+++ projects/storefrontlib/src/cms-components/product/product-list/container/product-list.component.spec.ts
@@ -22,7 +22,7 @@ import { ProductListComponent } from './product-list.component';
 import createSpy = jasmine.createSpy;
 import { ProductScrollComponent } from './product-scroll/product-scroll.component';
 import { InfiniteScrollModule } from 'ngx-infinite-scroll';
-import { ViewConfig } from '../../config/view-config';
+import { ViewConfig } from '../../../../view-config';
 
 @Component({
   selector: 'cx-star-rating',
